requirements.txt
CTkToolTip==0.8
customtkinter==5.2.2
numpy
openpyxl==3.2.0b1
pandas==2.2.3
pillow==11.1.0
PyMuPDF==1.26.4
tkinterdnd2==0.4.2
ttkwidgets==0.13.0
